requirements.txt
yt-dlp
demucs
openai-whisper
moviepy
pillow
numpy
torch
torchaudio
